requirements.txt
matplotlib==3.5.3
ml_collections==0.1.1
mlconfig==0.1.8
numpy==1.23.2
opencv_python==4.6.0.66
Pillow==9.3.0
scikit_image==0.19.3
scikit_learn==1.1.3
scipy==1.9.0
skimage==0.0
thop==0.1.1.post2209072238
torch==1.12.1+cu116
torchcam==0.3.2
torchvision==0.13.1+cu116
tqdm==4.64.1
